+++ examples/src/ammDeposit.js
@@ -4,12 +4,14 @@ const waitForDepositCreditedOnChain = require('./helpers/waitForDepositCreditedO
 
 const token1 = 'ETH'
 const token2 = 'USDT'
-const depositETHResponse = await dvf.deposit(token1, 0.1, starkPrivKey)
-const depositUSDTResponse = await dvf.deposit(token2, 1000, starkPrivKey)
+if (process.env.DEPOSIT_FIRST === 'true') {
+  const depositETHResponse = await dvf.deposit(token1, 0.1, starkPrivKey)
+  const depositUSDTResponse = await dvf.deposit(token2, 1000, starkPrivKey)
 
-if (process.env.WAIT_FOR_DEPOSIT_READY === 'true') {
-  await waitForDepositCreditedOnChain(dvf, depositETHResponse)
-  await waitForDepositCreditedOnChain(dvf, depositUSDTResponse)
+  if (process.env.WAIT_FOR_DEPOSIT_READY === 'true') {
+    await waitForDepositCreditedOnChain(dvf, depositETHResponse)
+    await waitForDepositCreditedOnChain(dvf, depositUSDTResponse)
+  }
 }
 
 const pool = `${token1}${token2}`
@@ -30,8 +32,8 @@ const ammFundingOrderData = await dvf.getAmmFundingOrderData({
 // This call signs the orders contained in the ammFundingOrderData before
 // posting them to Deversifi API. NOTE: if the orders are pre-signed, the
 // method will post them as is.
-const ammPostFundingOrderResponse = await dvf.postAmmFundingOrder(
-  ammFundingOrderData
+const ammPostFundingOrderResponse = await dvf.postAmmFundingOrders(
+  await dvf.applyFundingOrderDataSlippage(ammFundingOrderData, 0.05)
 )
 
 logExampleResult(ammPostFundingOrderResponse)