+++ src/api/depositV2.js
@@ -18,7 +18,8 @@ const schema = Joi.object({
   amount: Joi.bigNumber().greaterThan(0), // number or number string
   useProxiedContract: Joi.boolean().optional().default(false),
   permitParams: permitParamsSchema.optional(),
-  web3Options: Joi.object().optional() // For internal use (custom gas limits, etc)
+  web3Options: Joi.object().optional(), // For internal use (custom gas limits, etc)
+  referralId: Joi.string().optional()
 })
 
 const validateArg0 = validateWithJoi(schema)('INVALID_METHOD_ARGUMENT')({
@@ -29,7 +30,7 @@ const endpoint = '/v1/trading/deposits'
 const validationEndpoint = '/v1/trading/deposits-validate'
 
 module.exports = async (dvf, data, nonce, signature, txHashCb) => {
-  const { token, amount, useProxiedContract, web3Options, permitParams } = validateArg0(data)
+  const { token, amount, useProxiedContract, web3Options, permitParams, referralId } = validateArg0(data)
 
   const starkKey = dvf.config.starkKeyHex
 
@@ -85,8 +86,10 @@ module.exports = async (dvf, data, nonce, signature, txHashCb) => {
   const payload = {
     token,
     amount: quantisedAmount,
-    txHash: transactionHash
+    txHash: transactionHash,  
+    referralId
   }
+
   const httpDeposit = await post(dvf, endpoint, nonce, signature, payload)
 
   const onChainDeposit = await onChainDepositPromise