+++ src/api/migrationStampede/getPotValue.js
@@ -1,0 +1,6 @@
+const getGeneric = require('../../lib/dvf/get-generic')
+
+module.exports = async (dvf) => {
+  const endpoint = '/v1/trading/r/stampedePotValue'
+  return getGeneric(dvf, endpoint)
+}