+++ src/dvf.js
@@ -13,7 +13,7 @@ BigNumber.config({ EXPONENTIAL_AT: 1e9 })
  */
 module.exports = async (web3, userConfig = {}, sw) => {
   // binds all ./api methods into a fresh object, similar to creating an instance
-  let dvf = bind()
+  const dvf = bind()
   dvf.sw = sw
 
   // adds key-value storage and event emitting capabilities
@@ -23,10 +23,9 @@ module.exports = async (web3, userConfig = {}, sw) => {
   dvf.config = Object.assign({}, defaultConfig, userConfig)
 
   // ethfinex exchange config
-  const exchangeConf = dvf.config.autoLoadExchangeConf ?
-    await dvf.getConfig()
-    :
-    {}
+  const exchangeConf = dvf.config.autoLoadExchangeConf
+    ? await dvf.getConfig()
+    : {}
 
   // user config has priority
   dvf.config = Object.assign({}, defaultConfig, exchangeConf, userConfig)
@@ -60,7 +59,7 @@ module.exports = async (web3, userConfig = {}, sw) => {
   // If single web3 passed as parameter, assuming it is Ethereum by default
   } else {
     dvf.web3 = web3
-    dvf.web3PerChain = { ETHEREUM: web3 }
+    dvf.web3PerChain = { DEFAULT: web3, ETHEREUM: web3 }
   }
 
   if (!dvf.config.skipLoad) {
@@ -77,8 +76,7 @@ module.exports = async (web3, userConfig = {}, sw) => {
     // if (!dvf.get('account')) {
     //   console.warn('Please specify a valid account or account index')
     // }
-  }
-  else if (dvf.config.address) {
+  } else if (dvf.config.address) {
     dvf.set('account', dvf.config.address.toLowerCase())
   }