+++ src/lib/dvf/bindApi.js
@@ -12,8 +12,8 @@ module.exports = () => {
 
   // returns a function that will call api functions prepending dvf
   // as first argument
-  const compose = funk => {
-    return _partial(funk, dvf)
+  const compose = (funk, ...args) => {
+    return _partial(funk, dvf, ...args)
   }
 
   // dvf.account functions
@@ -22,7 +22,11 @@ module.exports = () => {
     tokenBalance: compose(require('../../api/account/tokenBalance')),
     select: compose(require('../../api/account/select')),
     getPermissions: compose(require('../../api/account/permissions').getPermissions),
-    setPermissions: compose(require('../../api/account/permissions').setPermissions)
+    setPermissions: compose(require('../../api/account/permissions').setPermissions),
+    getReferralId: compose(require('../../api/account/getReferralId')),
+    getRemainingSpins: compose(require('../../api/account/getRemainingSpins')),
+    getReferralRewards: compose(require('../../api/account/getReferralRewards')),
+    postReferralSpin: compose(require('../../api/account/postReferralSpin'))
   }
 
   dvf.stark = {
@@ -120,9 +124,24 @@ module.exports = () => {
     attachStarkProvider: compose(require('../../lib/wallet/attachStarkProvider'))
   }
 
-   // dvf.bitfinex functions
-   dvf.bitfinex = {
+  // dvf.bitfinex functions
+  dvf.bitfinex = {
     transfers: compose(require('../../api/bitfinex/transfers'))
+  }
+
+  // dvf.migrationStampede
+  dvf.migrationStampede = {
+    getStampedeConfig: compose(require('../../api/migrationStampede/getStampedeConfig')),
+    getMissionsConfig: compose(require('../../api/migrationStampede/getMissionsConfig')),
+    getPotValue: compose(require('../../api/migrationStampede/getPotValue')),
+    getUserMissions: compose(require('../../api/migrationStampede/getUserMissions')),
+    getUserReward: compose(require('../../api/migrationStampede/getUserReward'))
+  }
+
+  dvf.userVerification = {
+    setEmailOrPhone: compose(require('../../api/userVerification/setEmailOrPhone')),
+    verifyCode: compose(require('../../api/userVerification/verifyCode')),
+    isUserVerified: compose(require('../../api/userVerification/isUserVerified'))
   }
 
   // dvf.sign functions
@@ -131,7 +150,16 @@ module.exports = () => {
   dvf.sign.nonceSignature = compose(require('../../api/sign/nonceSignature'))
 
   dvf.postAuthenticated = compose(require('../../lib/dvf/post-authenticated'))
+
+  dvf.deleteAuthenticated = compose(require('../../lib/dvf/delete-authenticated'))
+
   dvf.getAuthenticated = compose(require('../../lib/dvf/get-authenticated'))
+
+  // Cancellable authenticated requests
+  dvf.request = ['get', 'post'].reduce((acc, method) => {
+    acc[method] = compose(require('../../lib/dvf/request'), method)
+    return acc
+  }, {})
 
   dvf.createOrderPayload = compose(require('../../lib/dvf/createOrderPayload'))
   dvf.createMarketOrderPayload = compose(require('../../lib/dvf/createMarketOrderPayload'))
@@ -199,7 +227,6 @@ module.exports = () => {
   dvf.getVaultIdFromServer = compose(require('../../api/getVaultIdFromServer'))
   dvf.getVaultIdAndStarkKey = compose(require('../../api/getVaultIdAndStarkKey'))
   dvf.register = compose(require('../../api/register'))
-  dvf.registerAndDeposit = compose(require('../../api/registerAndDeposit'))
   dvf.submitBuyOrder = compose(require('../../api/submitBuyOrder'))
   dvf.submitOrder = compose(require('../../api/submitOrder'))
   dvf.submitMarketOrder = compose(require('../../api/submitMarketOrder'))
@@ -219,6 +246,7 @@ module.exports = () => {
   dvf.postAmmFundingOrders = compose(require('../../api/amm/postAmmFundingOrders'))
   dvf.getAmmFundingOrders = compose(require('../../api/amm/getAmmFundingOrders'))
   dvf.getAmmFundingOrderData = compose(require('../../api/amm/getAmmFundingOrderData'))
+  dvf.applyFundingOrderDataSlippage = compose(require('../../api/amm/applyFundingOrderDataSlippage'))
   dvf.poolTVL = compose(require('../../api/amm/poolTVL'))
   dvf.poolTvlHistory = compose(require('../../api/amm/poolTvlHistory'))
   dvf.poolVolume24Hours = compose(require('../../api/amm/poolVolume24Hours'))
@@ -239,7 +267,8 @@ module.exports = () => {
     withdraw: compose(require('../../api/ledger/withdraw')),
     transfer: compose(require('../../api/ledger/transfer')),
     transferAndWithdraw: compose(require('../../api/ledger/transferAndWithdraw')),
-    transferUsingVaultIdAndStarkKey: compose(require('../../api/ledger/transferUsingVaultIdAndStarkKey'))
+    transferUsingVaultIdAndStarkKey: compose(require('../../api/ledger/transferUsingVaultIdAndStarkKey')),
+    signEIP712Data: compose(require('../../api/ledger/signEIP712Data'))
   }
   dvf.estimatedNextBatchTime = compose(require('../../api/estimatedNextBatchTime'))
   dvf.publicUserPermissions = compose(require('../../api/getPublicPermissions'))