+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.1.0 - 2022-06-30
+### Changed
+- Support fee on transferAndWithdraw
+
+## 4.0.0 - 2022-06-15
+### Changed
+- StarkEx v3 compatibility removed along with onchain registration functions
+
+## 3.1.9 - 2021-05-16
+### Changed
+- EIP712 signature helper for ledger devices
+
+## 3.1.8 - 2021-04-12
+### Changed
+- Add slippage reduction for funding stark orders
+
+## 3.1.7 - 2021-01-15
+### Changed
+- Store input "web3" as both ETHEREUM and DEFAULT web3 provider when not specified
+
 ## 3.1.6 - 2021-01-12
 ### Changed
 - Update bridge contract addresses conf location